base64 encoding/decoding actually works now

+++ src/base64.h
@@ -4,8 +4,11 @@
 #  define BASE64_H
 
 
-char* base64_encode( int *len, char *src, size_t sz );
-char* base64_decode( int *len, char *src, size_t sz );
+#include <stddef.h>
+
+
+char* base64_encode( size_t *len, char *src, size_t sz );
+char* base64_decode( size_t *len, char *src, size_t sz );
 
 
 #endif /* BASE64_H */
